requirements.txt
aiofiles==0.7.0
anyio==3.3.2
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
djangorestframework==3.13.1
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
market==0.0
MarkupSafe==2.0.1
numpy==1.22.3
otree==5.4.4
outcome==1.1.0
p5py==1.0.0
pep517==0.12.0
Pillow==9.0.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-decouple==3.6
python-multipart==0.0.5
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.22
sqlparse==0.4.2
starlette==0.16.0
threadpoolctl==3.1.0
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
uvicorn==0.15.0
websockets==10.0
Werkzeug==2.0.3
whitenoise==6.0.0
wsproto==1.0.0
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
zope.interface==5.4.0
